Clean up Dialyzer Sensor
